requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
mcp
aiohttp
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.3.0